Chore: 전체 모듈 패키지 구조 수정
