+++ module-api/build.gradle
@@ -23,6 +23,7 @@ repositories {
 
 dependencies {
     implementation project(':module-common')
+    implementation project(':module-domain')
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
