--- module-api/src/main/java/kernel/jdon/domain/faq/dto/request/UpdateFaqRequest.java
@@ -1,10 +1,0 @@
-package kernel.jdon.domain.faq.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class UpdateFaqRequest {
-	private Long faqId;
-	private String title;
-	private String content;
-}
