--- module-api/src/main/java/kernel/jdon/domain/faq/dto/response/DeleteFaqResponse.java
@@ -1,15 +1,0 @@
-package kernel.jdon.domain.faq.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class DeleteFaqResponse {
-
-	private Long faqId;
-
-	public static DeleteFaqResponse of(Long faqId) {
-		return new DeleteFaqResponse(faqId);
-	}
-}
