--- module-api/src/main/java/kernel/jdon/domain/faq/dto/response/UpdateFaqResponse.java
@@ -1,15 +1,0 @@
-package kernel.jdon.domain.faq.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UpdateFaqResponse {
-
-	private Long faqId;
-
-	public static UpdateFaqResponse of(Long faqId) {
-		return new UpdateFaqResponse(faqId);
-	}
-}
