--- module-api/src/main/java/kernel/jdon/domain/faq/entity/Faq.java
@@ -1,44 +1,0 @@
-package kernel.jdon.domain.faq.entity;
-
-import org.hibernate.annotations.DynamicUpdate;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import kernel.jdon.global.common.entity.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@DynamicUpdate
-@Table(name = "faq")
-public class Faq extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(name = "title", columnDefinition = "VARCHAR(255)", nullable = false)
-	private String title;
-
-	@Column(name = "content", columnDefinition = "TEXT", nullable = false)
-	private String content;
-
-	@Builder
-	public Faq(Long id, String title, String content) {
-		this.id = id;
-		this.title = title;
-		this.content = content;
-	}
-
-	public void update(String title, String content) {
-		this.title = title;
-		this.content = content;
-	}
-}
