--- module-api/src/main/java/kernel/jdon/domain/member/entity/Gender.java
@@ -1,17 +1,0 @@
-package kernel.jdon.domain.member.entity;
-
-public enum Gender {
-
-	MALE("남성"),
-	FEMALE("여성");
-
-	private final String gender;
-
-	Gender(String gender) {
-		this.gender = gender;
-	}
-
-	public String getGender() {
-		return gender;
-	}
-}
