--- module-api/src/main/java/kernel/jdon/domain/memberskill/entity/MemberSkill.java
@@ -1,29 +1,0 @@
-package kernel.jdon.domain.memberskill.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import kernel.jdon.domain.member.entity.Member;
-import kernel.jdon.domain.skill.entity.Skill;
-
-@Entity
-@Table(name = "member_skill")
-public class MemberSkill {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_id", columnDefinition = "BIGINT(50)")
-	private Member member;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "skill_id", columnDefinition = "BIGINT(50)")
-	private Skill skill;
-}
