--- module-api/src/main/java/kernel/jdon/domain/wantedskill/entity/WantedJdSkill.java
@@ -1,30 +1,0 @@
-package kernel.jdon.domain.wantedskill.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import kernel.jdon.domain.skill.entity.Skill;
-import kernel.jdon.domain.wanted.entity.WantedJd;
-
-@Entity
-@Table(name = "wanted_jd_skill")
-public class WantedJdSkill {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "skill_id", columnDefinition = "BIGINT(50)")
-	private Skill skill;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "wanted_jd_id", columnDefinition = "BIGINT(50)")
-	private WantedJd wantedJd;
-
-}
