+++ module-api/src/main/java/kernel/jdon/faq/repository/FaqRepository.java
@@ -1,0 +1,8 @@
+package kernel.jdon.faq.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import kernel.jdon.faq.domain.Faq;
+
+public interface FaqRepository extends JpaRepository<Faq, Long> {
+}
