--- module-api/src/main/java/kernel/jdon/global/common/entity/BaseEntity.java
@@ -1,24 +1,0 @@
-package kernel.jdon.global.common.entity;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-	@CreatedDate
-	@Column(name = "created_date", columnDefinition = "DATETIME", nullable = false, updatable = false)
-	private LocalDateTime createdDate;
-	@LastModifiedDate
-	@Column(name = "modified_date", columnDefinition = "DATETIME")
-	private LocalDateTime modifiedDate;
-}
