+++ module-api/src/test/java/kernel/jdon/moduleapi/domain/faq/service/FaqServiceTest.java
@@ -9,15 +9,15 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import kernel.jdon.domain.faq.dto.request.CreateFaqRequest;
-import kernel.jdon.domain.faq.dto.request.UpdateFaqRequest;
-import kernel.jdon.domain.faq.dto.response.CreateFaqResponse;
-import kernel.jdon.domain.faq.dto.response.DeleteFaqResponse;
-import kernel.jdon.domain.faq.dto.response.FindListFaqResponse;
-import kernel.jdon.domain.faq.dto.response.UpdateFaqResponse;
-import kernel.jdon.domain.faq.entity.Faq;
-import kernel.jdon.domain.faq.repository.FaqRepository;
-import kernel.jdon.domain.faq.service.FaqService;
+import kernel.jdon.faq.dto.request.CreateFaqRequest;
+import kernel.jdon.faq.dto.request.UpdateFaqRequest;
+import kernel.jdon.faq.dto.response.CreateFaqResponse;
+import kernel.jdon.faq.dto.response.DeleteFaqResponse;
+import kernel.jdon.faq.dto.response.FindListFaqResponse;
+import kernel.jdon.faq.dto.response.UpdateFaqResponse;
+import kernel.jdon.faq.domain.Faq;
+import kernel.jdon.faq.repository.FaqRepository;
+import kernel.jdon.faq.service.FaqService;
 
 @SpringBootTest
 public class FaqServiceTest {
