+++ module-common/src/main/java/kernel/jdon/dto/response/CommonResponse.java
@@ -1,0 +1,17 @@
+package kernel.jdon.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class CommonResponse<T> {
+
+	private T data;
+
+	public CommonResponse(T data) {
+		this.data = data;
+	}
+
+	public static <T> CommonResponse of(T data){
+		return new CommonResponse(data);
+	}
+}
