+++ module-domain/src/main/java/kernel/jdon/base/BaseEntity.java
@@ -1,0 +1,24 @@
+package kernel.jdon.base;
+
+import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseEntity {
+	@CreatedDate
+	@Column(name = "created_date", columnDefinition = "DATETIME", nullable = false, updatable = false)
+	private LocalDateTime createdDate;
+	@LastModifiedDate
+	@Column(name = "modified_date", columnDefinition = "DATETIME")
+	private LocalDateTime modifiedDate;
+}
