+++ module-domain/src/main/java/kernel/jdon/faq/dto/request/UpdateFaqRequest.java
@@ -1,0 +1,10 @@
+package kernel.jdon.faq.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class UpdateFaqRequest {
+	private Long faqId;
+	private String title;
+	private String content;
+}
