+++ module-domain/src/main/java/kernel/jdon/faq/dto/response/CreateFaqResponse.java
@@ -1,0 +1,15 @@
+package kernel.jdon.faq.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CreateFaqResponse {
+
+	private Long faqId;
+
+	public static CreateFaqResponse of(Long faqId) {
+		return new CreateFaqResponse(faqId);
+	}
+}
