+++ module-domain/src/main/java/kernel/jdon/faq/dto/response/FindFaqResponse.java
@@ -1,0 +1,21 @@
+package kernel.jdon.faq.dto.response;
+
+import kernel.jdon.faq.domain.Faq;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class FindFaqResponse {
+	private String title;
+	private String content;
+
+	public static FindFaqResponse of(Faq faq) {
+		return FindFaqResponse.builder()
+			.title(faq.getTitle())
+			.content(faq.getContent())
+			.build();
+	}
+}
