+++ module-domain/src/main/java/kernel/jdon/faq/dto/response/UpdateFaqResponse.java
@@ -1,0 +1,15 @@
+package kernel.jdon.faq.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UpdateFaqResponse {
+
+	private Long faqId;
+
+	public static UpdateFaqResponse of(Long faqId) {
+		return new UpdateFaqResponse(faqId);
+	}
+}
