+++ module-domain/src/main/java/kernel/jdon/favorite/domain/Favorite.java
@@ -1,0 +1,39 @@
+package kernel.jdon.favorite.domain;
+
+import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.CreatedDate;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import kernel.jdon.inflearn.domain.Inflearn;
+import kernel.jdon.member.domain.Member;
+
+@Entity
+@Table(name = "favorite")
+public class Favorite {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@CreatedDate
+	@Column(name = "created_date", columnDefinition = "DATETIME", nullable = false, updatable = false)
+	private LocalDateTime createdDate;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "member_id", columnDefinition = "BIGINT(50)")
+	private Member member;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "inflearn_id", columnDefinition = "BIGINT(50)")
+	private Inflearn inflearn;
+
+}
