+++ module-domain/src/main/java/kernel/jdon/member/domain/MemberRole.java
@@ -1,0 +1,17 @@
+package kernel.jdon.member.domain;
+
+public enum MemberRole {
+
+	ROLE_USER("사용자"),
+	ROLE_ADMIN("관리자");
+
+	private final String role;
+
+	MemberRole(String role) {
+		this.role = role;
+	}
+
+	public String getRole() {
+		return role;
+	}
+}
