+++ module-domain/src/main/java/kernel/jdon/memberskill/domain/MemberSkill.java
@@ -1,0 +1,29 @@
+package kernel.jdon.memberskill.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import kernel.jdon.member.domain.Member;
+import kernel.jdon.skill.domain.Skill;
+
+@Entity
+@Table(name = "member_skill")
+public class MemberSkill {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "member_id", columnDefinition = "BIGINT(50)")
+	private Member member;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "skill_id", columnDefinition = "BIGINT(50)")
+	private Skill skill;
+}
